Simplify showRebootMessage code
The existing code in showRebootMessage makes some assumptions
that are only valid in the a/b system world. In the new all-snaps
world of the future this will be no longer true and this branch
simplifies the code by making less assumptions.

There is a small UI change here, instead of:
 "Reboot to use the new ubuntu-core"
users will now see
 "Reboot to use ubuntu-core version 123"

If thats a problem we need to put some more work into this
branch to group releated snaps and see if the reboot-required
is triggered by an upgrade or a rollback.