Add "channel" information to `snap list`
This adds a new "Channel" column to the `snap list` output and also
adds a "(actual channel)" behind the name if the desired channel
and the actual channel are different. This can happen if e.g.
the user requests a snap from the edge channel, but the channel
is closed and the user gets a snap from the stable channel instead.

LP: #1591666