cmd: move snap-discard-ns to dedicated directory
This patch begins the dismantling of the monolithic cmd/snap-confine
directory so that eventually each executable can live in a dedicated
sub-directory of cmd/.

I also moved the manual page to be directly in the same place as the
corresponding executable.

Automake tweak: I needed to enable subdir-objects option so that this
builds correctly.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>