interfaces: convert systemd backend to new APIs
This patch changes the systemd backend and all associated interfaces
(currently that's only the GPIO interface) to the new interface APIs.

Since the systemd backend used the json-snippet hack (as a rough
prototype to the new API) the changes are actually smaller than in the
case of the mount backend. The JSON bridge is discarded and the GPIO
interface now uses just one method to convey the need for a custom
systemd service unit.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>