Move the snap-confine packaging into snapd
This solves the problem that users starting from snapd 2.0.10
(as shipped in Ubuntu 16.04.1) and then running "apt-et upgrade"
will not get an upgraded snapd because we have the new snap-confine
dependency in the later version of snapd (while the 2.0.10 version
had an ubuntu-core-launcher dependency).

By having it all in a single package "apt-get upgrade" will work
and snapd gets upgraded.