cmd: add /usr/local/* to PATH
Surprisingly /usr/local plays a role in the core snap. It is there where
the image build system splices additional files that don't come from a
typical deb. This includes the fake 'apt' script but most notably the
special 'xdg-open' executable that is necessary for snapd-xdg-open to
work.

The patch also expands the existing test that checks environment to
check PATH (and SNAP and HOME while I was a it it). One regression test
also needed changes as it was measuring the stability of PATH across
invocations of snap-confine.

Fixes: https://bugs.launchpad.net/snapcraft/+bug/1661023
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>