interfaces: use seccomp profile syntax readable by stable snap-confine
This patch reverts useful security hardening patches done by jdstrand in
order to unbreak people tracking the edge channel on core. Because
snap-confine is not re-executed yet we don't have a way to introduce new
seccomp syntax in a compatible way.

For the time being (before we allow snap-confine to run from core snap
directly) we need this to generate profiles that actually work.

This patch is designed to be reverted easily. It does not undo changes
to snap-confine that were made along the way as they are not the
problem.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>