cmd/snap-confine: ensure that hotfs is root owned.
On a system that never ran any snap before the /var/lib/snapd/hostfs
directory does not exist. When snap-confine is used it will create the
directory on demand but that directory will retain the group identity of
the user.

This patch fixes this issue by always chown'ing the said directory to be
owned by the root user and the root group.

Fixes: https://bugs.launchpad.net/snap-confine/+bug/1665004
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>