many: display kernel version in 'snap version'
This trivial branch displays the kernel version as reported by `uname`
the `release` field of `struct utsname`. There are no tests as this is
just a quick idea to explore.

Ideally we could ask the kernel "Are you certified" and we cold use that
as an answer but given that any kernel can lie to userspace this is not
so easy.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>