release: assume higher version of supported distros will still work (#2836)
The current code in release.go:os.ForceDevMode() uses a strict
version check to test if a new release supports non-devmode.

This is problematic as we always need to add new versions when
a derivate releases a new version. This branch uses VersionCompare()
to check the version and assumes that higher versions will continue
to work. It also assume somewhat sane versioning in VersionID. This
is the case so far and once we see expection we can always add code
for this.

The risk is of course that the distributions in higher version stop supporting
us. So this PR assumes that most derivates will not (i.e. it is less work to
add the exceptions than to track all the new versions).

This also adds peppermint os.