cmd/snap-update-ns: add function for sorting mount entries (#2923)
Mount entries can form a linked list that by itself is hard to sort. We
can resort to indirect sorting by allocating an array of pointers that
matches the length of the list. The array can be sorted with any of the
standard library routines (e.g. with qsort). Finally the linked list
must be re-written (all the next pointers as well as the first) pointer
to follow consecutive items in the intermediate array.