cmd/libsnap: constrain mount debug to -debug builds
In the non-debug build of snap-confine (the one that is setuid-root) we
will only print mount/umount debug data when we're about to die and
privileges are dropped.

In the -debug build we generate such messages (with all the added
complexity) depending on the runtime switch via the environment variable
(SNAP_CONFINE_DEBUG=yes)

The -debug build is not currently installed but we can eventually
install it (without any security elevation mechanism) and come up with a
way to use it (e.g. via a bind mount).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>