asserts: implement snap-developer type (#2302)
Initial snap-developer implementation.

The primary key is (snap-id, publisher-id) to allow an assertion to be emitted for a future publisher without disrupting things for the current publisher.

The assertion format is fully parsed (I think) and validated but there may be a need for additional validation, e.g. on date ranges.

There's no way to get at the developer date ranges atm. A method, e.g. DeveloperAllowed(developerID, when) bool, is probably best but there's no real need to add that until snapd needs to apply the assertion.