cmd/snap-update-ns: tweak everything to make testing work
This patch changes how the bootstrap code is enabled. I tried to make
several thins work (moving this to differnent packages, using multiple
consturctors with different prioritites and a flag variable) but it all
doesn't work because of how cgo linking works.

In absence of that, I just oped into a simple envirionment variable. I
use the value of SNAPD_INTERNAL and check for presence of
"x-switch-namespace=1," as a simple way to check if the bootstrap code
should work. It is easy to set this variable from snapd and we can now
unit test most of the boring C code easily.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>