cmd/snap-confine: set TMPDIR and TEMPDIR each time
We used to set TMPDIR and TEMPDIR and point them at /tmp when
constructing the mount namespace (the code was simply next to the code
that creates private /tmp) but after the move to per-snap persistent
namespaces the logic was unchanged. This means that the two variables
were set only the first time a given snap application was launched.

This patch corrects this by setting all environment in one place, right
next to the existing code that changes PATH.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>