many: model and expose interface meta-data.
This branch adds a new, optional, method that any snapd interface can
implement. The new method allows returning a structure with various
meta-data about the interface. For now only the Description field is
present.

This could be used for constructing various useful flows, like giving
interfaces proper descriptions, documentations URLs, information when it
was first added, possibly even static descriptions of desired
one-to-many connection semantics for slots.

In this experimental patch the common interface is extended to convey
the meta-data, a few interfaces are patched to document themselves and
the daemon and client APIs are extended to carry this information. The
command "snap interfaces -i IFACE" will now display the description of
the interface if one is available (try with network).

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>