tests/lib: introduce pkgdb helper library
All functions in the pkdb library will abstract different packaging
functions so that we have a common way to install/remove/update
packages in a distribution independent way.

This is required for later changes to add support for other distros
like Fedora or openSUSE.

The new helper functions are not used everywhere yet but follow up
changes will introduce their use everywhere so that everything is
abstracted through pkgdb functions.