many: stop "snap refresh $x --channel invalid" from working.
This splits LookupRefrsh out of store's ListRefresh, and makes
snapstate's Update use the new one. The difference is that
LookupRefresh can tell the difference between "this snap does not
exist [in the channel you asked about]" and "there are no updates for
this snap"; ListRefresh confounds the two situations, leading to some
hilarity.

Also includes some incidental refactoring of ListRefresh in order to
avoid DRY.