tests: fix for snapd-reexec test cheking for restart info on debug log
This change is to check the restart info a the debug log instaed of
checking that in the journalctl that sometimes does not log any data.

The prevoius check is not enough to validate the restart has been done
because journalctl used to show log information which already contains
the string "DEBUG: restarting into".

Error:
https://travis-ci.org/snapcore/snapd/builds/235819260