tests: fix timeout issue for test refresh core with hanging configure hook
This flaky test is failing in the autopkgtests execution. The failire is
caused because sometimes it is taking to much to generate the broken
core snap and the test is reaching the kill-timeout defined in 2
minutes.
The change makes the configure hangs during 60 seconds (enough time out
purpose). Then the kill time out is removed, avoiding any time out
issue.

ERROR:
https://objectstorage.prodstack4-5.canonical.com/v1/AUTH_77e2ada1e7a84929a74ba3b87153c0ac
/autopkgtest-xenial-snappy-dev-
image/xenial/i386/s/snapd/20170626_145348_a40d4@/log.gz