interfaces: add Repository.QueryInterfaces
This patch adds a new way to query interfaces present in the repository.
The mechanism is more flexible than prior attempts and in fact replaces
them in most cases. Subsequent patches will remove some unused code or
code that can be made unused through trivial calls to QueryInterfaces.

The input is a list of names and a set of options. The names decide which
interfaces are returned. As a special case, an empty list of names will
search all of the available interfaces. The options decide what data
is returned, such as documentation or plug and slot details. One of the
options also changes the scope of the search to just those interfaces
that have at least one connection.

The return value is a list of sorted InterfaceInfo objects with various
details filled in, as requested.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>