interfaces/builtin: don't sanitize interface type
This patch discards the code present in each Sanitize{Plug,Slot} that
was cross-checking iface.Name() with {plug,slot}.Interface.

This code was repetitive and served little use. There are four call
sites to Sanitize{Plug,Slot} in the whole codebase and they already
check the interface type up front.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>