apparmor,release: refactor apparmor probing / evaluation code
This patch is inspired by an idea from Michael Vogt. The probing code
now returns an apparmor.KernelSupport object which can be queried for
distinct facts: is apparmor enabled, is specific feature available.

The object can also be asked to evaluate overal support as required by
snapd.  This last operation matches the previous model where a tri-state
answer is provided (None, Partial, Full) as well as a textual summary
with more human-readable information ("this-and-that feature is
missing").

This will also allow specific interfaces to behave appropriately in
light of presence or absence of specific features.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>