interfaces/apparmor: add missing call to dirs.SetRootDir
This patch corrects a test that was failing on openSUSE. On Ubuntu it
was passing by accident because /var/cache/apparmor/$something did not
exist (the code treats file-not-found as an OK condition). On openSUSE
the permissions prevent non-root users from traversing
/var/cache/apparmor so the test failed with permission error.

In both cases what was missing was a call to dirs.SetRootDir(), to
isolate the test from the real system.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>