interfaces/network-{control,observe}: allow receiving kobject_uevent() messages
As part of the investigation for the livepatch regression and netlink socket
mediation, net/core/* from the Linux kernel sources shows that the kernel is
using kobject_uevent() for notifications. While we could force snaps to use
hardware-observe to view these, it makes sense in terms of completeness for
these interfaces and for compatibility with existing snaps to allow
'socket AF_NETLINK - NETLINK_KOBJECT_UEVENT' here as well.