snapstate: fix TestInstallWithoutCoreTwoSnapsWithFailureRunThrough
The TestInstallWithoutCoreTwoSnapsWithFailureRunThrough fails with
gccgo. The reason is that the taskrunner will run the prerequisites
task every 10ms - this will check if there is a pending core
link-snap task and if so it will trigger a retry. However the
retry is so short that link-snap never gets scheduled causing
a hang. By increasing the retry and decreasing the number of tests
this should be more reliable again.