snap-seccomp: run secondary-arch tests via gcc-multilib
When we run the seccomp tests via the in-kernel tests we currently
skip the secondary architecture (e.g. testing if 386 syscalls work
on amd64).

This patch tests this by building a 32bit binary on amd64 and
running the normal syscall tests. This needs gcc-multilib and
may not be available in other distros (but the test will fail
gracefully and just skip those tests).