data/completion: small tweak to snap completion snippet
previously, if you were completing from the middle of a command,
things might get weird. For example,

$ snap install test- --devmode
                    ^ you are here

if you hit tab there, without this change, you get a very helpful
“--devmode” replacing test-.

This is because we were just passing all the command off to go flags
for it to figure out, and it doesn't work that way.

Anyway, this fixes that.