snap-repair: fix tests when running as root
The tests are run as root under some test runners (like autopkgtest).
In this case os.Chmod(path, 0555) will not prevent writing to the
"path" because permissions are for lusers.

Skip these tests when run as root for now. We could make them work
with osutil.Chattr() but given that we test this already it is
probably not worth it.