Increase SNAPD_CONFIGURE_HOOK_TIMEOUT to 3 minutes to install real snaps
This is to give more time to real snaps to run the configure hook. For
example currently the juju snap is installing lxd as part of the
configure hooks and the current 30 seconds that we use for testing is
not enough.
This is used on a test which is set as manual and executed in a nightly
execution.