snap-{confine,seccomp}: make @unrestricted fully unrestricted
For @unrestricted we did load a "allow-all" bpf program so far.
However in LP: 1724697 the LXD team asks for fully unrestricted
mode (i.e. no filter at all) because they need to put their own
filter in place and they don't stack apparently.

The only downside/risk with this branch is that if snapd and
snap-confine get out of sync things will get messy as an old
snap-confine will try to load an invalid bpf program
(the program that reads "@unrestricted").