cmd/snap-update-ns: add designWritableMimic
This patch adds a helper function that inspects a directory and designs
changes that will create a "writable mimic" for that directory.

The writable mimic is done largely the same way we already do this in
the snap-confine source code (quirks.c) where we use a tmpfs that is
bind mounted over the original read-only place and a series of bind
mounts (a bind-mount farm essentially).

The solution presented here is more generic. The original directory is
"stashed" (bind mounted to a temporary directory), a tmpfs is mounted
and a series of bind mounts are created. In addition to supporting just
directories the code also supports creating files and even symlinks.

Files are bind-mounted just as directories but symlinks are created
directly, since the tmpfs is writable and symlinks cannot participate in
mounting.

The code only designs the necessary changes, the rest of the features
will come once prerequisite branches land.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>