interfaces: rename clashing mount entries
This patch changes how clashing mount entries are handled. As a simple
example consider a single app snap that has a content plug and two snaps
that provide a plugin via a content slot.

The interfaces are as follows:
 - app:plugins
 - plug1:plugin
 - plug2:plugin

If both plug1:plugin and plug2:plugin are connected to app:plugins and
the shared directory, in both cases, is "$SNAP/plugin", then the mount
entries for the app (assuming that target is "$SNAP/plugins") would be:

 /snap/plug1/1/plugin -> /snap/app/1/plugins/plugin
 /snap/plug1/1/plugin -> /snap/app/1/plugins/plugin

This arrangement is obviously invalid as the entries will overlap and
shadow. This is not desired.

With this patch the resulting mount entries are:

 /snap/plug1/1/plugin -> /snap/app/1/plugins/plugin
 /snap/plug1/1/plugin -> /snap/app/1/plugins/plugin-2

Note that the first entry is not renamed but subsequent entries get ever
increasing numbers, starting from -2.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>