daemon: Allow Polkit authorization to cancel changes.
We allow Polkit to authorize clients to install/remove snaps.
If a user wants to cancel this transaction, this requires root/macaroon access.

This allows a Polkit user to cancel any transaction - if we make Polkit access
more fine grained in the future we may want to check what the transaction type
is before picking the Polkit action id to use.