interfaces/udev: add a way to udev-tag devices for all apps/hooks
This patch introduces functionality similar to what other specification
types allow: the snippet is automatically added to all the apps and hooks
that affect the given plug or slot.

Udev is a bit special in that it allows arbitrary udev programs to be
injected (for instance, the network manager interface requires this)
but most interfaces don't need that generic functionality and instead
just want to tag a device so that all the apps associated with the
plug or slot have acess to it.

When we added hooks the udev backend was never updated to handle them
so we now noticed that hooks don't have any udev tagging at all.

With this new method we can use a much simpler API to do the right thing
while still retain the ability to add arbitrary free-form snippets
everywhere.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>