cmd/snap-update-ns: add XSnapdParentID
This patch adds another helper for determining the identifier of the
parent mount entry. The identifier can be used to associate mount
entries in a tree so that certain property can be easily determined.

snap-update-ns will soon synthesize new mount entries when trying to
perform a mount entry that cannot be completed because it affects a
read-only filesystem. To correctly track those for subsequent updates
such new synthesized entries will be "associated" with an entry in the
desired mount profile.

This will allow snap-update-ns to keep all synthesized mount entries as
long as their associated parent entry is still desired.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>