cmd: fix re-exec bug with classic confinement for host snapd < 2.28
Older version of snapd (before 2.28) did use the SNAPD_DID_REEXEC
env to check if they should re-exec or not. We need
to unset it after re-exec because the host snap tool may be old and
using this key.

So if e.g. the host has snapd 2.27 and snapd re-execs to 2.29 then
`snap run --shell classic-snap` will go into an environment where the
snapd 2.27 sees this key and stops re-execing and things break in
suble ways.

C.f.
https://forum.snapcraft.io/t/seccomp-error-calling-snap-from-another-classic-snap-on-core-candidate/2736/7