cmd/snap-update-ns: teach update logic to handle synthetic changes
This patch teaches the mount namespace update logic to cope with
synthetic changes. Such changes are are tagged with x-snapd.synethetic
field. Each syntethic change is tied via the x-snapd.parent-id field
to a non-synthetic change matching the x-snapd.id field.

As long as the non synthetic change is around all synthetic changes
that refer to it are preserved during updates. When the non-synthetic
change is no longer desired all the synthetic changes are undone in the
right order.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>