interfaces/browser-support: adjust base declaration for auto-connection (#4238)
During a review, pedronis discovered that the browser-support interface only
had a deny-connection constraint without a corresponding deny-auto-connection
constraint. Since constraints are evaluated independently of each other, this
meant that if a snap that used 'allow-sandbox: true' was manually approved in
the store without a snap declaration, then the interface would be
auto-connected. In practice, there were no snaps that were in the is state.
Existing snap declarations in the store for this interface have already been
adjusted to use 'allow-auto-connection: true' in addition to
'allow-connection: true'.