snapstate: ensure RefreshSchedule() gives accurate results (#4272)
The SnapManager.RefreshSchedule() did not give valid results until
ensureRefreshes() was run (which in turns ran checkRefreshSchedule).

This PR fixes this by refactoring the code so that RefreschSchedule
always reflects the current refresh schedule. It also avoids the
confusing side effect that "checkRefreshSchedule()" had and clarifies
that currentRefreshSchedule is actually the last known refresh
schedule (and it is only used to log if the option changed).