interfaces/browser-support: allow mknod() syscall for chromium nividiactl
Since snapd still uses SECCOMP_RET_KILL, we have added a seccomp workaround
rule to allow mknod on character devices because chromium unconditionally
performs a mknod() to create the /dev/nvidiactl device, regardless of if it
exists or not or if the process has CAP_MKNOD or not. Since we don't want to
actually grant the ability to create character devices, explicitly deny the
capability. When snapd uses SECCOMP_RET_ERRNO, we can remove these rules.

References:
https://forum.snapcraft.io/t/call-for-testing-chromium-62-0-3202-62/2569/46