snap: use the -no-fragments mksquashfs option
Use of the Squashfs fragments feature results in unpredictable snap
composition due to how fragments are gathered and compressed using a
separate thread in mksquashfs. Disable fragments entirely in `snap pack`
and tests that call mksquashfs directly so that the review tools can
perform validation of snaps and to be consistent with what snapcraft is
doing.

Fixes: https://launchpad.net/bugs/1576763
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>