overlord/snapstate: introduce refresh.timer core option
Introduce a new refresh.timer core option for keeping the refresh schedule in
the new format. This option has priority over refresh.schedule as far as the
'schedule' is concerned. The effect of setting refresh.schedule to 'managed' is
unchanged.

The default schedule is also expressed using the new format.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>