cmd/snap-update-ns: we don't want to bind mount symlinks
When planning a writable mimic we need to reconstruct the contents
of the original directory in a newly mounted tmpfs using bind mounts
and symlinks. Symlinks are represented as a fake mount entry with
"x-snapd.kind=symlink" field but it was also incorrectly specified as
a bind mount. While this is harmless let's not do that.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>