cmd/snap-update-ns: large refactor / update of unit tests
This patch complements earlier one that refactored changePerform.
This time all of the tests are updated but none of the non-test code
is touched. This was aimed at testing various black box behaviors
and getting coverage closer to 100%. In the two critical modules
(utils and change) this has been achieved.

The black box testing is designed to perform high-level operations
"attempt mount of such and such type and make it fail" which should
remain true regardless of further refactoring. The only part of the
test that bleeds through the black box is the list of syscalls made
by each test. This is tested explicitly to show what's going on,
mainly to security review.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>