snap: filter strace output
This ensures that only the part after `snap-exec` is visible to
the user and we skip all the snap-confine and snap-exec syscalls.

We filter on `execve("/snap/` and how all output after that. It
is not (quite) perfect as there will be a tiny bit of noise from
the go-routines of snap-exec that exit, like:

"""
  [pid  6770] <... futex resumed>)        = ?
  <... execve resumed> )                  = 0
"""

We could try to filter for that as well. However it seems britle.
Easy enough to add later if desired.