daemon: avoid panic'ing building an error response w/no snaps given
the (snapInstruction).errToResponse code in daemon/api.go assumed that
some of the errors were only possible when the instruction would have
at least one snap, and used that as part of the response message.

As the code has grown, with no checks for this assumption, it no
longer holds; we're seeing panics due to this.

This fixes that.