daemon: use either timer or schedule in refresh info
Adjust refresh info returned in sysinfo endpoint depending on the use of the
modern 'refresh.timer' or the legacy 'refresh.schedule' settings. When the new
refresh.timer setting is used only the 'timer' field will be sent. Similarly,
when the legacy setting is in effect, the 'timer' field will not be sent and
'schedule' will be send instead.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>