daemon: refactor snapFooMany helpers a little
The `snapInstallMany`, `snapUpdateMany`, and `snapRemoveMany` have the
same signature and were all returning a 4-uple of things. In working
to add another helper to the set I found I wanted to return a 5-uple,
so this change reworks them to instead return a struct (which makes it
a little easier to read already).