many: remove "content" argument from snaptest.MockSnap()
We currently pass a "content" parameter to MockSnap() which
describes what content the on-disk mocked snap file should
have. This is not used in any test, most tests set it to either
"" or "SNAP". So we just remove it from the API and set the
mocked snap content to "$REALNAME-$SNAPID-$REVNO" which has
the added benefit that the mock snap files on disk all have
different hashes which is useful when working with e.g.
the snap revision assertion.