wrappers: start multiple snap services at once
When a snap has a number of services we started them one by one. This does not
work well if services have an after/before ordering dependencies as there is no
way for system to enforce the ordering. Refactor the code so that snap services
are all started in one systemctl call.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>