snap: detect unsquashfs write failures
When unsquashfs fails to write files this is currently not reported
via the exit code. There is an upstream PR open about this:
  https://github.com/plougher/squashfs-tools/pull/46
However this is not merged yet. While waiting for this a workaround
is to look for "failed" in the unsquashfs output. This string is
not translated (nothing in unsquashfs is) and it is used pretty
consistently for error reporting.