snap: use custom unsquashfsStderrWriter for unsquashfs error detection
This addresses a review feedback suggestion from John in PR#4605
to keep the amount of error output we keep in memory smaller.

unsquahfs will potentially return an error for each file it tries to
unpack (e.g. on out of disk space) - so potentially a lot of output.

This PR creates a custom writer to minimize the impact. There is a
tradeoff here, the code is not as easy anymore, feedback welcome.

We might consider keeping the simplistic approach because we generally
do not unpack many files in snapd (bootloader asserts, kernel/initramfs)
so we shouldn't generate insane amount of output (currently at least).