debian/rules: workaround for https://github.com/golang/go/issues/23721
gccgo seems to have a bug where in some cases a goroutine never gets
to run. The issue can be worked around by having setting GOMAXPROCS to
something bigger than 1, so this change does that.

If do io copies on things that don't actually involve the operating
system we might bite the same issue outside of the tests, so if we
have gccgo-built binaries in production we probably want to set
GOMAXPROCS there as well.