devicestate: fix autopkgtest failure in TestDoRequestSerialErrorsOnNoHost
The autopkgtest environment uses a http_proxy during the tests. This
means that the invalid host in the test will not be looked up
directly but instead passed to the proxy. The proxy will return
a 503 which will be retried. So the error pattern is different.

For now the test is skipped when a http_proxy is set, we could
add a new test that checks the proxy behaviour.