interfaces/apparmor: always reload snap-confine profile
Because NFS workaround is handled separately the same correct
snap-confine profile may need to be re-loaded. This is because the
profile contains include directives and those load a second file that
has changed outside of the scope of EnsureDirState.

To counter that, always reload the profile by pretending it had changed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>