packaging/arch: sync with snapd/snapd-git from AUR
Both AUR packages snapd and snapd-git are mostly in sync with each other, with
the exception of some minor build logging and different `source` and so on.

Sync our in tree recipe with both snapd and snapd-git to the extent that makes
sense. Retain verbose build logs from snapd-git. Keep the check(), along with
running unit tests. Reorder install() section to match both AUR packages and
keep the diffs small.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>