packaging: create compat /lib/udev/snappy-app-dev file
Instead of a symlink from /lib/udev/snappy-app-dev we need to put
the real file there to support the following scenario. An older
snapd that expects /lib/udev/snappy-app-dev is run with a newer
core snap *but* with re-exec disabled. This will result in the
old snap-confine running /lib/udev/snappy-app-dev with apparmor
rules that will *only* allow to execute this file. So the previous
symlink approach is not good enough, it needs to be the real file
to allow this scenario.