jsonutil/puritan: drop duplicated checks
by the time puritan gets a json []byte it's already been checked for
validity, so we don't need to check check things that are invalid json.

Added explicit tests for the checks I removed, just in case.

I didn't set out to do this, but to try to increase coverage... and
all these ifs would always stay red :-)