osutil: fix fstab parser to allow for # in field values
This patch changes the fstab parser to allow the-hash-or-pound sign to
be used inside field values, as long as it is not the first character.

This specifically allows us to parse this mount entry:

    mhddfs#/mnt/hdd1,/mnt/hdd2,/mnt/hdd3 /mnt/virtual fuse defaults,allow_other 0 0

The parser no longer strips off first encountered '#' sign. Instead
comments are differentiated into full comments and inline comments. Full
comments have '#' as the first non-whitespace character of a given line
and are stripped out by the fstab file parser and don't reach the fstab
entry parser. Inline comments are stripped after fstab-entry field
separation and must contain '#' as the first character of a field. This
combination allows '#' to be safely used inside individual fields, as
long as it is not the first character.

Fixes: https://bugs.launchpad.net/snappy/+bug/1760841
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>