cmd/snap-confine: ignore missing cgroups in snap-device-helper
Don't attempt to modify cgroups that don't exist.

The snap-device-helper script is invoked by udev as a result of a
snapd-generated udev rule. When a device is tagged to be visible to a
given snap the tag is picked up by snap-confine and placed into the
device cgroup on application startup.

Long running applications have their cgroups modified by the very same
udev rule, because apart from tagging the rule runs the
snap-device-helper script to alter the cgroup.

Prior to this patch the modification was done regardless if the cgroup
existed or not, resulting in repeated error messages:

     Apr 08 15:23:32 vougeot systemd-udevd[919]: Process
     '/usr/lib/snapd/snap-device-helper add snap_vlc_vlc
     /devices/pci0000:00/0000:00:02.0/drm/card0 226:0' failed with exit code
     2.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1762182
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>