tests: add check for OOM error after each test
We see errors in autopkgtest on bionic/i386 currently. Laney was
kind enough to do a debug run and he noticed that the OOM killer
was invoked during the run. The most likely reason for this is:
https://forum.snapcraft.io/t/oom-for-interfaces-many-on-bionic-i386/4101

But the error pattern in autopkgtest is confusing, when it happens
it will simply hang in "journalctl --sync" and eventually timeout.

This PR makes the error explicit.