interfaces/builtin: allow abstract unix sockets, handle Ubuntu 14.04 special case
Even though everyone seems to be using /run/systemd/notify, NOTIFY_SOCKET can
also be an abstract Unix socket. Extend the interface to support such sockets
and generate proper AppArmor rules.

Ubuntu 14.04 is a special case. It does use abstract sockets, but the path
changes with each reboot. Since we are aware of this behavior, add it as a
special case.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>