interfaces/apparmor: fix incorrect apparmor profile glob
This patch fixes a bug where a glob describing apparmor profiles for a
snap would match unrelated snap that happens to be a proper prefix of
the name of the snap in question.

What was missing was a dot between the snap name and the glob that
describes all applications and hooks. In the broken version it was
snap.$SNAP_NAME*.

For example, setting up security for "snapcraft" would nuke the profiles
of "snapcraft-pr".

This patch also includes small helper computes the name of the
snap-update-ns profile for a given snap.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>